+++ refreshed/src/main/java/com/neuandroid/news/model/SourcesList.java
@@ -1,4 +1,4 @@
-package com.neuandroid.news;
+package com.neuandroid.news.model;
 
 /**
  * Created by max on 19/07/2017.
