+++ refreshed/src/main/java/com/neuandroid/xkcd/util/IAsyncTaskListener.java
@@ -1,4 +1,4 @@
-package com.neuandroid.xkcd;
+package com.neuandroid.xkcd.util;
 
 import java.io.Serializable;
